Final Update
StringStuff.java is a method to detect the characteristics of a string. The program is capable of detecting if the string is a palindrome, contains vowel, remove duplicates from a string, and output the odd and even characters in a string. 

When running the main program, StringStuff.java tests each function in the program and shows an example of them working.